Cargo.toml: remove oldtime from chrono features because of RUSTSEC-2020-0071
valico does not seem to be using any of the chrono oldtime feature bits
so let's disable it to not bring in the time crate and not be vulnerable
to RUSTSEC-2020-0071.

Right now, chrono does not support time >= 0.2 which is not vulnerable
to the stated issue. Once it does, valico might get a dependency version
dump for chrono and this patch can be reverted.

Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>